kernel: package MHI bus support
ath11k requires MHI bus support for the PCI card support, so package
it as kernel module so it can be added as a dependency to the PCI ath11k
driver module.

Currently, only 5.15 kernel is supported as backporting fixes would be
required for it to function under 5.10 at all.

Signed-off-by: Robert Marko <robimarko@gmail.com>
Signed-off-by: boos4721 <3.1415926535boos@gmail.com>